Bump timestamp to u64

+++ packages/prop-house-contracts/contracts/starknet-cairo-v1/src/common/lib/traits.cairo
@@ -8,7 +8,7 @@ trait IExecutionStrategy {
 #[abi]
 trait IVotingStrategy {
     fn get_voting_power(
-        timestamp: u32,
+        timestamp: u64,
         voter_address: ContractAddress,
         params: Array<felt252>,
         user_params: Array<felt252>,
